--- prometheus.yml
@@ -1,9 +1,0 @@
-scrape_configs:
-  # The job name is added as a label `job=<job_name>` to any timeseries scraped from this config.
-  - job_name: 'tekuLighthouse5'
-
-    # Override the global default and scrape targets from this job every 5 seconds.
-    scrape_interval: 5s
-    scheme: http
-    static_configs:
-        - targets: ['ec2-18-209-46-236.compute-1.amazonaws.com:8008']
