+++ storage/src/test/java/tech/pegasys/artemis/storage/RecentChainDataTest.java
@@ -14,7 +14,6 @@
 package tech.pegasys.artemis.storage;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.argThat;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
@@ -23,9 +22,6 @@ import static org.mockito.Mockito.when;
 
 import com.google.common.eventbus.EventBus;
 import com.google.common.primitives.UnsignedLong;
-import java.util.Optional;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.atomic.AtomicBoolean;
 import org.apache.tuweni.bytes.Bytes32;
 import org.junit.jupiter.api.Test;
 import tech.pegasys.artemis.datastructures.blocks.BeaconBlock;
@@ -34,13 +30,10 @@ import tech.pegasys.artemis.datastructures.state.Checkpoint;
 import tech.pegasys.artemis.datastructures.state.MutableBeaconState;
 import tech.pegasys.artemis.datastructures.util.DataStructureUtil;
 import tech.pegasys.artemis.storage.Store.Transaction;
-import tech.pegasys.artemis.storage.api.StorageUpdateChannel;
 import tech.pegasys.artemis.storage.events.FinalizedCheckpointEvent;
-import tech.pegasys.artemis.util.async.SafeFuture;
-import tech.pegasys.artemis.util.async.StubAsyncRunner;
 import tech.pegasys.artemis.util.config.Constants;
 
-class ChainStorageClientTest {
+class RecentChainDataTest {
 
   private static final BeaconState INITIAL_STATE =
       new DataStructureUtil(3).randomBeaconState(UnsignedLong.ZERO);
@@ -50,101 +43,11 @@ class ChainStorageClientTest {
 
   private final DataStructureUtil dataStructureUtil = new DataStructureUtil();
   private final EventBus eventBus = mock(EventBus.class);
-  private final StorageUpdateChannel storageUpdateChannel = mock(StorageUpdateChannel.class);
-  private final StubAsyncRunner asyncRunner = new StubAsyncRunner();
   private final Store store = mock(Store.class);
   private final RecentChainData storageClient =
       MemoryOnlyRecentChainData.createWithStore(eventBus, store);
   private final RecentChainData preGenesisStorageClient =
       MemoryOnlyRecentChainData.create(eventBus);
-
-  @Test
-  public void storageBackedClient_storeInitializeViaGetStoreRequest()
-      throws ExecutionException, InterruptedException {
-    SafeFuture<Optional<Store>> storeRequestFuture = new SafeFuture<>();
-    when(storageUpdateChannel.onStoreRequest()).thenReturn(storeRequestFuture);
-
-    final EventBus eventBus = new EventBus();
-    final SafeFuture<RecentChainData> client =
-        StorageBackedRecentChainData.create(asyncRunner, storageUpdateChannel, eventBus);
-
-    // We should have posted a request to get the store from storage
-    verify(storageUpdateChannel).onStoreRequest();
-
-    // Client shouldn't be initialized yet
-    assertThat(client).isNotDone();
-
-    // Post a store response to complete initialization
-    final Store genesisStore = Store.get_genesis_store(INITIAL_STATE);
-    storeRequestFuture.complete(Optional.of(genesisStore));
-    assertThat(client).isCompleted();
-    assertStoreInitialized(client.get());
-    assertStoreIsSet(client.get());
-    assertThat(client.get().getStore()).isEqualTo(genesisStore);
-  }
-
-  @Test
-  public void storageBackedClient_storeInitializeViaNewGenesisState()
-      throws ExecutionException, InterruptedException {
-    SafeFuture<Optional<Store>> storeRequestFuture = new SafeFuture<>();
-    when(storageUpdateChannel.onStoreRequest()).thenReturn(storeRequestFuture);
-
-    final EventBus eventBus = new EventBus();
-    final SafeFuture<RecentChainData> client =
-        StorageBackedRecentChainData.create(asyncRunner, storageUpdateChannel, eventBus);
-
-    // We should have posted a request to get the store from storage
-    verify(storageUpdateChannel).onStoreRequest();
-    // Client shouldn't be initialized yet
-    assertThat(client).isNotDone();
-
-    // Post a store event to complete initialization
-    storeRequestFuture.complete(Optional.empty());
-    assertThat(client).isCompleted();
-    assertStoreNotInitialized(client.get());
-    assertThat(client.get().getStore()).isNull();
-
-    // Now set the genesis state
-    final Store genesisStore = Store.get_genesis_store(INITIAL_STATE);
-    client.get().initializeFromGenesis(INITIAL_STATE);
-    assertStoreInitialized(client.get());
-    assertStoreIsSet(client.get());
-    assertThat(client.get().getStore()).isEqualTo(genesisStore);
-  }
-
-  // TODO: figure out how to simulate a timeout and uncomment this test
-  //  @Test
-  //  public void storageBackedClient_storeInitializeViaGetStoreRequestAfterTimeout()
-  //      throws ExecutionException, InterruptedException {
-  //    SafeFuture<Optional<Store>> storeRequestFuture = new SafeFuture<>();
-  //    when(storageUpdateChannel.onStoreRequest())
-  //            .thenReturn(storeRequestFuture);
-  //
-  //    final EventBus eventBus = new EventBus();
-  //    final SafeFuture<RecentChainData> client =
-  //            StorageBackedRecentChainData
-  //                    .create(asyncRunner, storageUpdateChannel, eventBus);
-  //
-  //    // We should have posted a request to get the store from storage
-  //    verify(storageUpdateChannel).onStoreRequest();
-  //
-  //
-  //    // Client shouldn't be initialized yet
-  //    assertThat(client).isNotDone();
-  //
-  //    // Wait for the initial request to timeout
-  //    Thread.sleep(3000);
-  //
-  //    assertThat(asyncRunner.countDelayedActions()).isEqualTo(1);
-  //
-  //    // Now set the genesis state
-  //    final Store genesisStore = Store.get_genesis_store(INITIAL_STATE);
-  //    storeRequestFuture.complete(Optional.of(genesisStore));
-  //    assertThat(client).isCompleted();
-  //    assertStoreInitialized(client.get());
-  //    assertStoreIsSet(client.get());
-  //    assertThat(client.get().getStore()).isEqualTo(genesisStore);
-  //  }
 
   @Test
   public void initialize_setupInitialState() {
@@ -374,27 +277,6 @@ class ChainStorageClientTest {
     final Checkpoint currentCheckpoint =
         preGenesisStorageClient.getStore().getFinalizedCheckpoint();
     assertThat(currentCheckpoint).isEqualTo(originalCheckpoint);
-  }
-
-  private void assertStoreInitialized(final RecentChainData client) {
-    final AtomicBoolean initialized = new AtomicBoolean(false);
-    client.subscribeStoreInitialized(() -> initialized.set(true));
-    assertThat(initialized).isTrue();
-  }
-
-  private void assertStoreNotInitialized(final RecentChainData client) {
-    final AtomicBoolean initialized = new AtomicBoolean(false);
-    client.subscribeStoreInitialized(() -> initialized.set(true));
-    assertThat(initialized).isFalse();
-  }
-
-  private void assertStoreIsSet(final RecentChainData client) {
-    assertThat(client.getStore()).isNotNull();
-
-    // With a store set, we shouldn't be allowed to overwrite the store by setting the genesis state
-    assertThatThrownBy(() -> client.initializeFromGenesis(INITIAL_STATE))
-        .isInstanceOf(IllegalStateException.class)
-        .hasMessageContaining("Failed to set genesis state: store has already been initialized");
   }
 
   private boolean isFinalizedCheckpointEvent(final Object obj) {
